R4

+++ Lexico.cs
@@ -1,5 +1,5 @@
 //BRIONES ALMAGUER CINTHYA CRISTINA
-using 
+using
 System.IO;
 
 namespace Generador
@@ -13,7 +13,7 @@ namespace Generador
         const int F = -1;
         const int E = -2;
         protected int linea, posicion;
-       
+
         int[,] TRAND = new int[,]
         {
             {0,1,5,3,4,5},
@@ -37,7 +37,7 @@ namespace Generador
             programa = new StreamWriter("C:\\Generico\\Programa.cs");
             programa.AutoFlush = true;
             log.WriteLine("Archivo: c.gram");
-            
+
             log.WriteLine(DateTime.Now);
             if (existencia == true)
             {
@@ -52,24 +52,36 @@ namespace Generador
         {
             linea = 1;
             string pathLog = Path.ChangeExtension(nombre, ".log");
-            log = new StreamWriter(pathLog); 
+            /*GetExtension(String):
+Devuelve la extensión (incluido el punto ".") de la cadena de ruta de acceso especificada.*/
+            string existeGram = Path.GetExtension(nombre);
+            //Console.WriteLine(existeGram);
+            log = new StreamWriter(pathLog);
             log.AutoFlush = true;
-            //LEGUAJE
-            lenguaje = new StreamWriter("C:\\Generico\\Lenguaje.cs");
-            lenguaje.AutoFlush = true;
-            //PROGRAMA
-            programa = new StreamWriter("C:\\Generico\\Programa.cs");
-            programa.AutoFlush = true;
-            log.WriteLine("Archivo: "+nombre);
-            log.WriteLine("Fecha: " +DateTime.Now);
-            if (File.Exists(nombre))
+            if (existeGram == ".gram")
             {
-                archivo = new StreamReader(nombre);
+                //LEGUAJE
+                lenguaje = new StreamWriter("C:\\Generico\\Lenguaje.cs");
+                lenguaje.AutoFlush = true;
+                //PROGRAMA
+                programa = new StreamWriter("C:\\Generico\\Programa.cs");
+                programa.AutoFlush = true;
+                log.WriteLine("Archivo: " + nombre);
+                log.WriteLine("Fecha: " + DateTime.Now);
+                if (File.Exists(nombre))
+                {
+                    archivo = new StreamReader(nombre);
+                }
+                else
+                {
+                    throw new Error("Error: El archivo " + Path.GetFileName(nombre) + " no existe ", log);
+                }
             }
             else
             {
-                throw new Error("Error: El archivo " +Path.GetFileName(nombre)+ " no existe ", log);
+                throw new Error("Error: En la extencion del archivo", log);
             }
+
         }
         public void cerrar()
         {
@@ -77,11 +89,11 @@ namespace Generador
             log.Close();
             lenguaje.Close();
             programa.Close();
-        }       
+        }
 
         private void clasifica(int estado)
         {
-            switch(estado)
+            switch (estado)
             {
                 case 1:
                     setClasificacion(Tipos.ST);
@@ -115,25 +127,25 @@ namespace Generador
             else
                 return 5;
         }
-        public void NextToken() 
+        public void NextToken()
         {
-            string buffer = "";           
-            char c;      
+            string buffer = "";
+            char c;
             int estado = 0;
-            while(estado >= 0)
+            while (estado >= 0)
             {
                 c = (char)archivo.Peek(); //Funcion de transicion
-                estado = TRAND[estado,columna(c)];
+                estado = TRAND[estado, columna(c)];
                 clasifica(estado);
                 if (estado >= 0)
                 {
                     archivo.Read();
                     posicion++;
-                    if(c == '\n')
+                    if (c == '\n')
                     {
                         linea++;
                     }
-                    if (estado >0)
+                    if (estado > 0)
                     {
                         buffer += c;
                     }
@@ -143,10 +155,10 @@ namespace Generador
                     }
                 }
             }
-            setContenido(buffer); 
-            if(estado == E)
+            setContenido(buffer);
+            if (estado == E)
             {
-                throw new Error("Error lexico: No definido en linea: "+linea, log);
+                throw new Error("Error lexico: No definido en linea: " + linea, log);
             }
             if (!FinArchivo())
             {
